refactor: rename `runtime_base_registers` to `runtime`
Too long and `registers` repetition.
